Move GraphiQL E2E into its own workspace
