+++ functions/graphql.ts
@@ -9,8 +9,8 @@ import {
   processRequest,
   type ProcessRequestResult,
 } from 'graphql-helix';
-import { createSchema } from '../packages/graphiql/test/schema.js';
-import { createExecute } from '../packages/graphiql/test/execute.js';
+import { createSchema } from '../packages/graphiql-e2e/test/schema.js';
+import { createExecute } from '../packages/graphiql-e2e/test/execute.js';
 
 type HandlerOptions = {
   schema: graphql.GraphQLSchema;
