+++ package.json
@@ -7,7 +7,7 @@
     "packages": [
       "packages/*",
       "functions",
-      "packages/graphiql/test",
+      "packages/graphiql-e2e/test",
       "examples/monaco-graphql-webpack",
       "examples/monaco-graphql-nextjs",
       "examples/monaco-graphql-react-vite",
@@ -24,12 +24,12 @@
   },
   "scripts": {
     "types:check": "turbo run types:check && tsgo --noEmit -p scripts/tsconfig.json",
-    "dev:graphiql": "turbo run dev --filter=graphiql...",
+    "dev:graphiql": "turbo run dev --filter=graphiql-e2e...",
     "dev:example-nextjs": "turbo run dev --filter=example-graphiql-nextjs...",
     "dev:example-vite": "turbo run dev --filter=example-graphiql-vite...",
     "storybook": "yarn workspace @graphiql/react storybook",
     "build-storybook": "yarn workspace @graphiql/react build-storybook",
-    "build:graphiql": "turbo run build --filter=graphiql...",
+    "build:graphiql": "turbo run build --filter=graphiql-e2e...",
     "build": "yarn build-clean && yarn build:tsc && yarn build:nontsc",
     "build-bundles": "yarn prebuild-bundles && yarn wsrun:noexamples --stages build-bundles",
     "build-bundles-clean": "rimraf '{packages,examples}/**/{bundle,cdn,webpack}'",
@@ -42,8 +42,8 @@
     "watch": "yarn build:watch",
     "watch-vscode": "yarn build:tsc && yarn workspace vscode-graphql compile",
     "watch-vscode-exec": "yarn workspace vscode-graphql-execution compile",
-    "cypress-open": "concurrently 'yarn dev:graphiql' 'yarn workspace graphiql cypress-open'",
-    "e2e": "yarn workspace graphiql e2e",
+    "cypress-open": "concurrently 'yarn dev:graphiql' 'yarn workspace graphiql-e2e cypress-open'",
+    "e2e": "yarn workspace graphiql-e2e e2e",
     "oxlint": "oxlint --max-warnings=0 --ignore-path .eslintignore",
     "fix": "yarn lint-fix && yarn format && yarn svgo",
     "license-check": "jsgl --local .",
