+++ packages/graphiql-e2e/cypress.config.ts
