+++ packages/graphiql-e2e/cypress/e2e/docs.cy.ts
