+++ packages/graphiql-e2e/cypress/e2e/errors.cy.ts
