+++ packages/graphiql-e2e/cypress/e2e/headers.cy.ts
