+++ packages/graphiql-e2e/cypress/e2e/keyboard.cy.ts
