+++ packages/graphiql-e2e/cypress/e2e/merge-fragments.cy.ts
