+++ packages/graphiql-e2e/cypress/e2e/operation-cursor.cy.ts
