+++ packages/graphiql-e2e/cypress/e2e/prettify.cy.ts
