+++ packages/graphiql-e2e/cypress/e2e/query-builder.cy.ts
