+++ packages/graphiql-e2e/cypress/e2e/tabs.cy.ts
