+++ packages/graphiql-e2e/cypress/env.d.ts
