+++ packages/graphiql-e2e/cypress/fixtures/bad-schema.json
