+++ packages/graphiql-e2e/cypress/fixtures/fixtures.ts
