+++ packages/graphiql-e2e/cypress/plugins/index.ts
