+++ packages/graphiql-e2e/cypress/tsconfig.json
