+++ packages/graphiql-e2e/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "graphiql-e2e",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "license": "MIT",
+  "scripts": {
+    "types:check": "tsgo --noEmit",
+    "build": "vite build --config vite.e2e.config.mts",
+    "build-demo": "copy index.html ../graphiql/ && mkdirp ../graphiql/dist/e2e && cp -R dist/e2e/. ../graphiql/dist/e2e",
+    "cypress-open": "cypress open --browser electron",
+    "dev": "concurrently 'cross-env PORT=8080 node test/e2e-server' vite",
+    "e2e": "yarn e2e-server 'cypress run'",
+    "test:a11y:update": "CYPRESS_A11Y_UPDATE_BASELINE=1 yarn e2e-server 'cypress run --spec cypress/e2e/a11y.cy.ts'",
+    "e2e-server": "start-server-and-test 'cross-env PORT=8080 node test/e2e-server' 'http-get://localhost:8080/graphql?query={test { id }}'"
+  },
+  "dependencies": {
+    "@graphiql/react": "1.0.0-beta.1",
+    "@graphiql/toolkit": "1.0.0-beta.0",
+    "graphiql": "6.0.0-beta.1",
+    "graphql": "^16.14.2",
+    "graphql-ws": "^5.5.5",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0"
+  },
+  "devDependencies": {
+    "@typescript/native-preview": "^7.0.0-dev",
+    "@vitejs/plugin-react": "^4.4.1",
+    "axe-core": "^4",
+    "babel-plugin-react-compiler": "19.1.0-rc.3",
+    "concurrently": "^7.0.0",
+    "copy": "^0.3.2",
+    "cross-env": "^7.0.2",
+    "cypress": "^13.13.2",
+    "cypress-axe": "^1",
+    "cypress-real-events": "^1.15.0",
+    "mkdirp": "^1.0.4",
+    "start-server-and-test": "^1.10.11",
+    "typescript": "^5.8.0",
+    "vite": "^6.4.3"
+  }
+}
