+++ packages/graphiql-e2e/test/package.json
