+++ packages/graphiql/package.json
@@ -24,8 +24,7 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
-    "dist",
-    "!dist/e2e/**"
+    "dist"
   ],
   "exports": {
     "./package.json": "./package.json",
@@ -39,12 +38,7 @@
   },
   "scripts": {
     "types:check": "tsgo --noEmit",
-    "build": "vite build && vite build --config vite.e2e.config.mts",
-    "cypress-open": "cypress open --browser electron",
-    "dev": "concurrently 'cross-env PORT=8080 node test/e2e-server' vite",
-    "e2e": "yarn e2e-server 'cypress run'",
-    "test:a11y:update": "CYPRESS_A11Y_UPDATE_BASELINE=1 yarn e2e-server 'cypress run --spec cypress/e2e/a11y.cy.ts'",
-    "e2e-server": "start-server-and-test 'cross-env PORT=8080 node test/e2e-server' 'http-get://localhost:8080/graphql?query={test { id }}'",
+    "build": "vite build",
     "test": "vitest run"
   },
   "dependencies": {
@@ -68,19 +62,11 @@
     "@testing-library/react": "^16.3.0",
     "@vitejs/plugin-react": "^4.4.1",
     "@vitest/web-worker": "^4.1.6",
-    "axe-core": "^4",
     "babel-plugin-react-compiler": "19.1.0-rc.3",
-    "concurrently": "^7.0.0",
-    "cross-env": "^7.0.2",
-    "cypress": "^13.13.2",
-    "cypress-axe": "^1",
-    "cypress-real-events": "^1.15.0",
     "graphql": "^16.14.2",
-    "graphql-ws": "^5.5.5",
     "lightningcss": "^1.29.3",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "start-server-and-test": "^1.10.11",
     "typescript": "^5.8.0",
     "vite": "^6.4.3",
     "vite-plugin-dts": "^4.5.3",
