--- packages/graphql-language-service-server/src/__tests__/__utils__/runServer.js
@@ -1,1 +1,0 @@
-exports.default = require('../../../../graphiql/test/e2e-server.js');
